feat: create routes files and change general code structure
